requirements.txt
opencv-python
deepface
librosa
numpy
tk
sounddevice
pillow